CompanyImport: Unset events only for bulk imports
It makes a difference only if there are many records being imported and
in other times it actually might be useful to be able to see the logged
queries.

When the verbosity level of the command is set to 'debug', then don't
unset event dispatcher. It lets to collect the query logs with the
expense of progressively increased memory usage.

Signed-off-by: Reinis Danne <rei4dan@gmail.com>